requirements.txt
amqp==5.1.1
antlr4-python3-runtime==4.11.1
antlr4-tools==0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
awscli==1.27.79
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
bleach==6.0.0
boto3==1.26.79
botocore==1.29.79
Brotli==1.0.9
bytecode==0.14.2
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickhouse-connect==0.6.3
colorama==0.4.4
comm==0.1.2
ConfigArgParse==1.5.3
confluent-kafka==2.1.1
cryptography==39.0.1
ddsketch==2.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==1.2.0
dj-email-url==1.0.6
dj-rest-auth==2.2.7
Django==4.1.4
django-allauth==0.52.0
django-appconf==1.0.5
django-cache-url==3.4.4
django-celery-beat==2.4.0
django-celery-email==3.0.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-prometheus==2.2.0
django-redis==5.2.0
django-ses==3.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.16
envier==0.4.0
environs==9.5.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.2.0
fqdn==1.5.1
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
importlib-metadata==6.0.1
install-jdk==0.3.0
ipykernel==6.20.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kombu==5.2.4
locust==2.15.1
lz4==4.3.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.5
mypy==0.931
mypy-extensions==0.4.3
mypy-protobuf==3.4.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
opentelemetry-api==1.18.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.6
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
rsa==4.7.2
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
types-protobuf==4.21.0.2
typing_extensions==4.4.0
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.3.4
widgetsnbextension==4.0.5
wrapt==1.15.0
xmltodict==0.13.0
zipp==3.11.0
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
django-redis==5.2.0
django-clickhouse-backend==1.0.3
datadog==0.46.0
datadog-api-client==2.15.0
newrelic-api==1.0.6

